requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
matplotlib>=3.5.0
pandas>=1.3.0
transformers>=4.30.0
Pillow>=9.0.0
requests>=2.25.0
argparse>=1.4.0
opencv-python>=4.5.0
tqdm>=4.65.0
pycocotools>=2.0.0
pyyaml>=6.0
